Added bitmap-based shader
Added the possibilty to choose a bitmap based shader to draw with. This
included a re-structuring of how the BrushShader handled added points as
many assumptions related to only using GL_LINES no longer held. The
resulting re-factoring should be easier to extend and easier to use,
notably, the demo interface to the shader only passes down new points to
draw and the line-begin / line-end logic is moved into the BrushShader
where it belongs. This also simplifes adding prediction points.

Added a ByteBuffer management class to handle all the OpenGL buffers
needed for drawing textures.

One included bitmap-based brush is included (spraypaint).